Changed: Add empty controllers, contarcts and file repository+service

+++ ZaylandShop.ProductCatalog.Storage/Repositories/Repository.cs
@@ -15,7 +15,7 @@ namespace ZaylandShop.ProductCatalog.Storage.Repositories
 			_dbSet = _dbContext.Set<TEntity>();
 		}
 
-		public async Task<TEntity> GetByIdAsync(int id)
+		public async Task<TEntity> GetByIdAsync(long id)
 		{
 			return await _dbSet.FindAsync(id);
 		}
